Merge TP before migration

+++ src/Importer/Parsing/SoftTree.cpp
@@ -10,11 +10,14 @@ namespace Parsing {
 
 QString Modifier::ToString( ModifierEnum modifier )
 {
+#include <leathers/push>
+#include <leathers/exit-time-destructors>
 	static const std::map<ModifierEnum, QString> enumToStr {
 		{ PUBLIC, "public" },
 		{ PROTECTED, "protected" },
 		{ PRIVATE, "private" }
 	};
+#include <leathers/pop>
 
 	auto found = enumToStr.find( modifier );
 	return found != enumToStr.end() ? found->second : QString();
@@ -64,7 +67,7 @@ bool Method::HasResult() const
 
 uint Method::GetLineOfCodes() const
 {
-	return content.count( '\n' );
+	return static_cast<unsigned int>( content.count( '\n' ) );
 }
 
 // Class
